Hide database ownership behind persistence client
